- Avanzando pues

+++ evenxusreports/frontend/cargarreporte.php
@@ -102,17 +102,24 @@ function CargarReporte($Reporte) {
         print "Instalando reporte...<br><br>";
         $BP->ActualizarPantalla();
         $NombreReporte = basename($Reporte, ".zip");
-        include("../upload/".$NombreReporte."/install.php");
-        // Limpieza final
-        print "Limpieza de temporales...<br><br>";
-        $BP->ActualizarPantalla();
-        BorrarCarpeta("../upload/$NombreReporte");
-        unlink ("../upload/$Reporte");
-        print "Reporte instalado correctamente...redirigiendo al reporte<br><br>";
-        $BP->ActualizarPantalla();
-        sleep(10);
-        $Redirigir=DOL_MAIN_URL_ROOT."/evenxusreports/frontend/".$NombreFiltros;
-        print "<script language='javascript'>window.location='$Redirigir'</script>;";
+        $id_menu_superior=ObtenerIDMenuSuperior("evenxusreports");
+        // Obtiene menu superior(Menu evenxusreports) e instala
+        if ($id_menu_superior>-1) {
+            include("../upload/".$NombreReporte."/install.php");
+            // Limpieza final
+            print "Limpieza de temporales...<br><br>";
+            $BP->ActualizarPantalla();
+            BorrarCarpeta("../upload/$NombreReporte");
+            unlink ("../upload/$Reporte");
+            print "Reporte instalado correctamente...redirigiendo al reporte<br><br>";
+            $BP->ActualizarPantalla();
+            sleep(10);
+            $Redirigir=DOL_MAIN_URL_ROOT."/evenxusreports/frontend/".$NombreFiltros;
+            print "<script language='javascript'>window.location='$Redirigir'</script>;";
+        }
+        else {
+            print "Error en la instalacion del modulo. No se puede encontrar el menu principal para instalar submenus.";            
+        }
     } else {
         print "Error al descomprimir el instalador. Es posible que no sea un ZIP o este corrupto.";
     }
